+++ src/mediawiki.ts
@@ -19,6 +19,7 @@ import {
 import {MediaWiki} from './token.js';
 import {
 	getCompletions,
+	getExtTags,
 } from './util.js';
 import type {
 	StreamParser,
@@ -212,8 +213,7 @@ export class FullMediaWiki extends MediaWiki {
 				};
 			}
 			mt = context.matchBefore(/<\/?[a-z\d]*$/iu);
-			const extTags = [...types].filter(type => type.startsWith('mw-tag-'))
-				.map(s => s.slice(7));
+			const extTags = getExtTags([...types]);
 			if (mt && (explicit || mt.to - mt.from > 1)) {
 				const validFor = /^[a-z\d]*$/iu;
 				if (mt.text[1] === '/') {